dev/bingoapp/src/main/java/be/cegeka/devopscourse/demo/ItemService.java
package be.cegeka.devopscourse.demo;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ItemService {

    @Autowired
    private ItemRepository itemRepository;

    public void saveItem(Item item) {
        itemRepository.save(item);
    }

    public List<Item> findAllItems() {
        return itemRepository.findAll();
    }

    public void delete(String id) {
        itemRepository.delete(id);
    }
}
